backend/requirements.txt
Django==3.2
django-cors-headers==3.13.0
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.0.0
PyYAML==6.0
psycopg2-binary==2.9.3
python-dotenv
hashids==1.3.1